+++ shared/src/commonMain/kotlin/co/edu/unicauca/dopaminah/App.kt
@@ -6,13 +6,18 @@ import co.edu.unicauca.dopaminah.ui.navigation.DopamiNahApp
 import co.edu.unicauca.dopaminah.ui.navigation.LocalPermissionState
 import co.edu.unicauca.dopaminah.ui.navigation.PermissionState
 import co.edu.unicauca.dopaminah.ui.screens.dashboard.viewmodel.DashboardViewModel
+import co.edu.unicauca.dopaminah.ui.screens.stats.viewmodel.StatsViewModel
 
 @Composable
 fun App(
     permissionState: PermissionState = PermissionState(),
-    dashboardViewModel: DashboardViewModel? = null
+    dashboardViewModel: DashboardViewModel? = null,
+    statsViewModel: StatsViewModel? = null
 ) {
     CompositionLocalProvider(LocalPermissionState provides permissionState) {
-        DopamiNahApp(dashboardViewModel = dashboardViewModel)
+        DopamiNahApp(
+            dashboardViewModel = dashboardViewModel,
+            statsViewModel = statsViewModel
+        )
     }
 }